use fragments instead of activities
